move benchmark test to a separate package
